[Cursor] Refactor backend app.py into modules
This commit refactors the monolithic app.py file into a modular architecture:

- Created models package for Pydantic schemas
- Created services package for business logic
- Created routes package for API endpoints
- Simplified main app.py to focus on initialization

This modular structure improves maintainability, readability, and makes future
feature development easier.

Refs #11
