+++ claude-tooling/app/api/app.py
@@ -4,34 +4,16 @@ FastAPI application to integrate Claude 3.7 with local tools.
 """
 
 import os
-import json
 import logging
-import time
-import datetime
-import mimetypes
-from pathlib import Path
-from typing import Dict, List, Any, Optional, Union
-from fastapi import FastAPI, HTTPException, BackgroundTasks, Request, Depends
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
-from fastapi.responses import RedirectResponse, JSONResponse, FileResponse, HTMLResponse
-from pydantic import BaseModel
+from fastapi.responses import RedirectResponse
 import anthropic
 from dotenv import load_dotenv
-import uuid
-import asyncio
 
-# Import local tool modules
-from ..tools.tool_wrapper import (
-    TOOL_DEFINITIONS,
-    process_tool_calls,
-    format_tool_results_for_claude,
-    set_conversation_root_dir,
-    get_conversation_root_dir
-)
-
-# Load environment variables
-load_dotenv()
+# Import routes
+from .routes import chat_router, conversation_router, files_router
 
 # Configure logging
 logging.basicConfig(
@@ -39,6 +21,9 @@ logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 )
 logger = logging.getLogger(__name__)
+
+# Load environment variables
+load_dotenv()
 
 # Initialize FastAPI app
 app = FastAPI(
@@ -62,593 +47,19 @@ try:
         api_key=os.environ.get("ANTHROPIC_API_KEY")
     )
     logger.info("Anthropic client initialized successfully")
+
+    # Inject the client into the chat router
+    from .routes.chat import client as chat_client
+    chat_client = client
+
 except Exception as e:
     logger.error(f"Failed to initialize Anthropic client: {str(e)}")
     raise
 
-# Pydantic models for request/response
-class Message(BaseModel):
-    role: str
-    content: List[Dict[str, Any]] = []
-    
-class ToolUse(BaseModel):
-    id: str
-    type: str
-    name: str
-    input: Dict[str, Any]
-    
-class ToolOutput(BaseModel):
-    tool_use_id: str
-    content: str
-    
-class UserRequest(BaseModel):
-    messages: List[Message]
-    max_tokens: int = 4096
-    temperature: float = 0.7
-    thinking_mode: bool = True
-    thinking_budget_tokens: int = 2000
-    auto_execute_tools: bool = True
-    
-class UserResponse(BaseModel):
-    message: Message
-    conversation_id: Optional[str] = None
-    tool_calls: List[Dict[str, Any]] = []
-    thinking: Optional[str] = None
-
-# Store conversations
-conversations = {}
-# Store conversation root directories
-conversation_root_dirs = {}
-
-# 存储正在执行的自动工具调用任务状态
-auto_execute_tasks = {}
-
-# Function to create a conversation root directory
-def create_conversation_root_dir(conversation_id: str) -> str:
-    """
-    Create a directory for a conversation based on timestamp.
-    
-    Args:
-        conversation_id: The conversation ID
-        
-    Returns:
-        The path to the created directory
-    """
-    timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-    root_dir = os.path.join("runs", timestamp)
-    
-    # Create the directory if it doesn't exist
-    os.makedirs(root_dir, exist_ok=True)
-    
-    # Store the root directory for this conversation
-    conversation_root_dirs[conversation_id] = root_dir
-    
-    logger.info(f"Created root directory for conversation {conversation_id}: {root_dir}")
-    
-    # Update the tool wrapper with the conversation root directory
-    set_conversation_root_dir(conversation_id, root_dir)
-    
-    return root_dir
-
-# 添加一个新的路由参数
-class UserRequest(BaseModel):
-    messages: List[Message]
-    max_tokens: int = 4096
-    temperature: float = 0.7
-    thinking_mode: bool = True
-    thinking_budget_tokens: int = 2000
-    auto_execute_tools: bool = True  # New parameter to control automatic tool execution
-
-# 添加新的自动工具执行函数
-async def auto_execute_tool_calls(tool_calls: List[Dict[str, Any]], conversation_id: str = None) -> List[Dict[str, Any]]:
-    """
-    Automatically execute tool calls and return results.
-    
-    Args:
-        tool_calls: List of tool calls to execute
-        conversation_id: Optional conversation ID for context
-    """
-    try:
-        logger.info(f"Automatically executing {len(tool_calls)} tool calls")
-        
-        # Use the existing process_tool_calls function to handle tool calls
-        # Since process_tool_calls may not be an async function, ensure proper handling
-        tool_results = process_tool_calls(tool_calls, conversation_id)
-        
-        # Return formatted tool results
-        logger.info(f"Tool execution completed, {len(tool_results)} results obtained")
-        return tool_results
-    
-    except Exception as e:
-        logger.error(f"Error executing tool calls: {str(e)}")
-        # Return error results
-        error_results = []
-        for tool_call in tool_calls:
-            error_results.append({
-                "tool_use_id": tool_call.get("id"),
-                "content": json.dumps({
-                    "status": "error",
-                    "message": f"Error executing tool calls: {str(e)}"
-                }, ensure_ascii=False)
-            })
-        return error_results
-
-# Modified chat route to support automatic tool execution
-@app.post("/api/chat", response_model=UserResponse)
-async def chat(request: UserRequest, background_tasks: BackgroundTasks, conversation_id: Optional[str] = None):
-    """
-    Process a chat request to Claude 3.7, handling tool calls.
-    """
-    try:
-        logger.info("Received chat request")
-        
-        # 首先检查URL参数中是否提供了conversation_id
-        if not conversation_id:
-            # 如果URL参数中没有提供，尝试从请求体的系统消息中提取
-            for message in request.messages:
-                if message.role == "system" and any(content.get("conversation_id") for content in message.content if isinstance(content, dict)):
-                    conversation_id = next((content.get("conversation_id") for content in message.content if isinstance(content, dict) and "conversation_id" in content), None)
-                    break
-        
-        # 如果仍未找到，创建新的conversation_id
-        if not conversation_id:
-            conversation_id = f"conv_{int(time.time())}"
-            logger.info(f"Created new conversation ID: {conversation_id}")
-            
-            # Create a root directory for this conversation
-            create_conversation_root_dir(conversation_id)
-        else:
-            logger.info(f"Using existing conversation ID: {conversation_id}")
-            
-            # Create root directory if it doesn't exist for this conversation
-            if conversation_id not in conversation_root_dirs:
-                create_conversation_root_dir(conversation_id)
-        
-        # Prepare messages for Claude API
-        api_messages = []
-        
-        # Extract system message content for Claude API
-        system_content = "You are running in a headless environment. When generating code that creates visualizations or outputs, DO NOT use interactive elements like plt.show(), figure.show(), or display()."
-        
-        for msg in request.messages:
-            # Format message content for the API
-            content_list = msg.content
-            if isinstance(content_list, list):
-                api_messages.append({
-                    "role": msg.role,
-                    "content": content_list
-                })
-            else:
-                # If content is not a list, make it a list with a text item
-                api_messages.append({
-                    "role": msg.role,
-                    "content": [{"type": "text", "text": content_list}]
-                })
-        
-        # Make the API call to Claude
-        logger.info("Sending request to Claude API")
-        
-        # Prepare thinking parameter if thinking mode is enabled
-        thinking_param = None
-        temperature_param = request.temperature
-        if request.thinking_mode:
-            thinking_param = {
-                "type": "enabled",
-                "budget_tokens": request.thinking_budget_tokens
-            }
-            # When thinking is enabled, temperature must be 1
-            temperature_param = 1.0
-            logger.info("Thinking mode enabled, setting temperature to 1.0")
-        
-        response = client.messages.create(
-            model="claude-3-7-sonnet-20250219",
-            system=system_content,
-            messages=api_messages,
-            max_tokens=request.max_tokens,
-            temperature=temperature_param,
-            tools=TOOL_DEFINITIONS,
-            thinking=thinking_param,
-        )
-        
-        logger.info("Received response from Claude API")
-        logger.info(f"Response type: {type(response)}")
-        
-        # Extract thinking if available
-        thinking = None
-        if hasattr(response, 'thinking'):
-            thinking = response.thinking
-            logger.info(f"Found thinking: {thinking}")
-        
-        # Convert response content to dictionaries
-        try:
-            # First try to dump the model to a dict
-            if hasattr(response, 'model_dump'):
-                response_dict = response.model_dump()
-                logger.info(f"Successfully dumped response model")
-                raw_content = response_dict.get('content', [])
-            else:
-                # Fallback: try to access content directly
-                raw_content = response.content
-                logger.info(f"Accessed content directly: {type(raw_content)}")
-            
-            # Convert content items to dictionaries
-            content = []
-            for item in raw_content:
-                if hasattr(item, 'model_dump'):
-                    content.append(item.model_dump())
-                elif hasattr(item, '__dict__'):
-                    content.append(item.__dict__)
-                else:
-                    # If it's already a dict or compatible
-                    content.append(dict(item))
-                    
-            logger.info(f"Processed content items: {len(content)}")
-        except Exception as e:
-            logger.error(f"Error converting response content: {str(e)}")
-            raise HTTPException(status_code=500, detail=f"Error processing response: {str(e)}")
-        
-        # Process tool calls
-        tool_calls = []
-        try:
-            for item in content:
-                if item.get('type') == 'tool_use':
-                    tool_calls.append({
-                        "id": item.get('id'),
-                        "name": item.get('name'),
-                        "input": item.get('input', {})
-                    })
-            
-            if tool_calls:
-                logger.info(f"Found {len(tool_calls)} tool calls")
-        except Exception as e:
-            logger.error(f"Error processing tool calls: {str(e)}")
-            
-        # Update the conversation history
-        if conversation_id not in conversations:
-            conversations[conversation_id] = []
-        
-        # Add user message and assistant response to conversation history
-        conversations[conversation_id].extend([
-            {"role": "user", "content": request.messages[-1].content},
-            {"role": "assistant", "content": content}
-        ])
-        
-        # If auto-execute tools is enabled and there are tool calls, process them in the background
-        if request.auto_execute_tools and tool_calls:
-            logger.info("Automatic tool execution enabled, starting tool task")
-            
-            # Create a background task to handle tool calls and continue conversation
-            background_tasks.add_task(process_tool_calls_and_continue, tool_calls, conversation_id, request.max_tokens, request.thinking_mode, request.thinking_budget_tokens, request.auto_execute_tools)
-            
-            # Don't include tool calls in the response when auto-execution is enabled
-            # The frontend will get them through polling instead
-            tool_calls_for_response = []
-        else:
-            # For manual execution, include tool calls in the response
-            tool_calls_for_response = tool_calls
-        
-        # Prepare the response
-        return UserResponse(
-            message=Message(role="assistant", content=content),
-            conversation_id=conversation_id,
-            tool_calls=tool_calls_for_response,
-            thinking=thinking
-        )
-        
-    except Exception as e:
-        logger.error(f"Error processing chat request: {str(e)}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-# Modified process_tool_calls_and_continue function to add interruption checks
-async def process_tool_calls_and_continue(
-    tool_calls: List[Dict[str, Any]], 
-    conversation_id: str, 
-    max_tokens: int, 
-    thinking_mode: bool, 
-    thinking_budget_tokens: int,
-    auto_execute_tools: bool
-):
-    """
-    Process tool calls, get results, and continue conversation with Claude until all tool calls are completed.
-    """
-    try:
-        if not tool_calls:
-            logger.info("No tool calls to process")
-            return
-            
-        # Get conversation history
-        if conversation_id not in conversations:
-            logger.error(f"Conversation ID not found: {conversation_id}")
-            return
-        
-        # Initialize or check task status
-        if conversation_id not in auto_execute_tasks:
-            auto_execute_tasks[conversation_id] = "running"
-        
-        # Check if cancelled
-        if auto_execute_tasks.get(conversation_id) == "cancelled":
-            logger.info(f"Automatic execution of conversation {conversation_id} cancelled")
-            return
-            
-        history = conversations[conversation_id]
-        
-        # Automatically execute tool calls and get results
-        try:
-            tool_results = await auto_execute_tool_calls(tool_calls, conversation_id)
-        except Exception as e:
-            logger.error(f"Error executing tool calls: {str(e)}")
-            # Add error message to history
-            history.append({
-                "role": "system",
-                "content": [{"type": "text", "text": f"Error executing tool calls: {str(e)}"}]
-            })
-            auto_execute_tasks[conversation_id] = "error"
-            return
-        
-        # Check if cancelled
-        if auto_execute_tasks.get(conversation_id) == "cancelled":
-            logger.info(f"Automatic execution of conversation {conversation_id} cancelled")
-            return
-        
-        # Format tool results as expected by Claude API
-        tool_result_blocks = format_tool_results_for_claude(tool_results)
-        
-        # Add tool results to conversation history
-        history.append({
-            "role": "user",
-            "content": tool_result_blocks
-        })
-        
-        # Prepare thinking parameter
-        thinking_param = None
-        temperature_param = 0.7
-        if thinking_mode:
-            thinking_param = {
-                "type": "enabled",
-                "budget_tokens": thinking_budget_tokens
-            }
-            temperature_param = 1.0
-        
-        # Check if cancelled
-        if auto_execute_tasks.get(conversation_id) == "cancelled":
-            logger.info(f"Automatic execution of conversation {conversation_id} cancelled")
-            return
-            
-        # Call Claude API to continue conversation
-        logger.info(f"Continuing conversation with Claude, conversation ID: {conversation_id}")
-        try:
-            response = client.messages.create(
-                model="claude-3-7-sonnet-20250219",
-                system="You are running in a headless environment. When generating code that creates visualizations or outputs:\n"\
-                      "1. DO NOT use interactive elements like plt.show(), figure.show(), or display()\n"\
-                      "2. Instead, save outputs to files (e.g., plt.savefig('output.png'))\n"\
-                      "3. For Python plots, use matplotlib's savefig() method\n"\
-                      "4. For Jupyter-style outputs, write to files instead\n"\
-                      "5. Always provide complete, self-contained code that can run without user interaction\n"\
-                      "6. Assume your code runs in a script context, not an interactive notebook",
-                messages=history,
-                max_tokens=max_tokens,
-                temperature=temperature_param,
-                tools=TOOL_DEFINITIONS,
-                thinking=thinking_param,
-            )
-        except Exception as e:
-            logger.error(f"Error calling Claude API: {str(e)}")
-            # Add error message to history
-            history.append({
-                "role": "system",
-                "content": [{"type": "text", "text": f"Error calling Claude API: {str(e)}"}]
-            })
-            auto_execute_tasks[conversation_id] = "error"
-            return
-        
-        # Process response
-        try:
-            # Extract response content
-            if hasattr(response, 'model_dump'):
-                response_dict = response.model_dump()
-                raw_content = response_dict.get('content', [])
-            else:
-                raw_content = response.content
-                
-            # Convert content items to dictionaries
-            content = []
-            for item in raw_content:
-                if hasattr(item, 'model_dump'):
-                    content.append(item.model_dump())
-                elif hasattr(item, '__dict__'):
-                    content.append(item.__dict__)
-                else:
-                    content.append(dict(item))
-                    
-            # Add assistant response to conversation history
-            history.append({"role": "assistant", "content": content})
-            
-            # Check if cancelled
-            if auto_execute_tasks.get(conversation_id) == "cancelled":
-                logger.info(f"Automatic execution of conversation {conversation_id} cancelled")
-                return
-                
-            # Extract new tool calls
-            new_tool_calls = []
-            for item in content:
-                if item.get('type') == 'tool_use':
-                    new_tool_calls.append({
-                        "id": item.get('id'),
-                        "name": item.get('name'),
-                        "input": item.get('input', {})
-                    })
-                    
-            # If there are new tool calls and automatic execution is enabled, recursively process
-            if auto_execute_tools and new_tool_calls:
-                logger.info(f"Found {len(new_tool_calls)} new tool calls, continuing processing")
-                await process_tool_calls_and_continue(
-                    new_tool_calls, 
-                    conversation_id, 
-                    max_tokens, 
-                    thinking_mode, 
-                    thinking_budget_tokens,
-                    auto_execute_tools
-                )
-                
-        except Exception as e:
-            logger.error(f"Error processing Claude response: {str(e)}")
-            # Add error message to history
-            history.append({
-                "role": "system",
-                "content": [{"type": "text", "text": f"Error processing Claude response: {str(e)}"}]
-            })
-            auto_execute_tasks[conversation_id] = "error"
-            
-    except Exception as e:
-        logger.error(f"Error processing tool calls and continuing conversation: {str(e)}")
-        # Add error message to conversation history
-        if conversation_id in conversations:
-            conversations[conversation_id].append({
-                "role": "system",
-                "content": [{"type": "text", "text": f"Error processing tool calls and continuing conversation: {str(e)}"}]
-            })
-        auto_execute_tasks[conversation_id] = "error"
-    finally:
-        # Clean up state after task completion
-        if conversation_id in auto_execute_tasks and auto_execute_tasks[conversation_id] not in ["cancelled", "error"]:
-            auto_execute_tasks[conversation_id] = "completed"
-
-# Modified /api/tool-results endpoint to support automatic tool execution flag
-@app.post("/api/tool-results", response_model=UserResponse)
-async def submit_tool_results(
-    tool_output: ToolOutput,
-    background_tasks: BackgroundTasks,
-    conversation_id: str,
-    auto_execute_tools: bool = True
-):
-    """
-    Process tool outputs and continue the conversation with Claude.
-    """
-    try:
-        logger.info(f"Received tool output for conversation {conversation_id}")
-        
-        if conversation_id not in conversations:
-            raise HTTPException(status_code=404, detail="Conversation not found")
-        
-        # Get conversation history
-        history = conversations[conversation_id]
-        
-        # Add tool output to conversation history
-        history.append({
-            "role": "user",
-            "content": [
-                {
-                    "type": "tool_result",
-                    "tool_use_id": tool_output.tool_use_id,
-                    "content": tool_output.content
-                }
-            ]
-        })
-        
-        # Make the API call to continue the conversation
-        logger.info("Sending tool result to Claude API")
-        response = client.messages.create(
-            model="claude-3-7-sonnet-20250219",
-            system="You are running in a headless environment. When generating code that creates visualizations or outputs:\n"\
-                  "1. DO NOT use interactive elements like plt.show(), figure.show(), or display()\n"\
-                  "2. Instead, save outputs to files (e.g., plt.savefig('output.png'))\n"\
-                  "3. For Python plots, use matplotlib's savefig() method\n"\
-                  "4. For Jupyter-style outputs, write to files instead\n"\
-                  "5. Always provide complete, self-contained code that can run without user interaction\n"\
-                  "6. Assume your code runs in a script context, not an interactive notebook",
-            messages=history,
-            max_tokens=4096,
-            temperature=1.0,  # Must be 1.0 when thinking is enabled
-            tools=TOOL_DEFINITIONS,
-            thinking={
-                "type": "enabled",
-                "budget_tokens": 2000
-            },
-        )
-        
-        logger.info("Received response from Claude API")
-        logger.info(f"Response type: {type(response)}")
-        
-        # Convert response content to dictionaries
-        try:
-            # First try to dump the model to a dict
-            if hasattr(response, 'model_dump'):
-                response_dict = response.model_dump()
-                logger.info(f"Successfully dumped response model")
-                raw_content = response_dict.get('content', [])
-            else:
-                # Fallback: try to access content directly
-                raw_content = response.content
-                logger.info(f"Accessed content directly: {type(raw_content)}")
-            
-            # Convert content items to dictionaries
-            content = []
-            for item in raw_content:
-                if hasattr(item, 'model_dump'):
-                    content.append(item.model_dump())
-                elif hasattr(item, '__dict__'):
-                    content.append(item.__dict__)
-                else:
-                    # If it's already a dict or compatible
-                    content.append(dict(item))
-                    
-            logger.info(f"Processed content items: {len(content)}")
-        except Exception as e:
-            logger.error(f"Error converting response content: {str(e)}")
-            raise HTTPException(status_code=500, detail=f"Error processing response: {str(e)}")
-        
-        # Process tool calls
-        tool_calls = []
-        try:
-            for item in content:
-                if item.get('type') == 'tool_use':
-                    tool_calls.append({
-                        "id": item.get('id'),
-                        "name": item.get('name'),
-                        "input": item.get('input', {})
-                    })
-            
-            if tool_calls:
-                logger.info(f"Found {len(tool_calls)} tool calls")
-        except Exception as e:
-            logger.error(f"Error processing tool calls: {str(e)}")
-        
-        # Add assistant response to conversation history
-        history.append({"role": "assistant", "content": content})
-        
-        # If automatic execution is enabled and there are tool calls, start background task to handle
-        if auto_execute_tools and tool_calls:
-            logger.info("Automatic tool execution enabled, starting tool task")
-            
-            # Create a background task to handle tool calls and continue conversation
-            background_tasks.add_task(
-                process_tool_calls_and_continue, 
-                tool_calls, 
-                conversation_id, 
-                4096,  # max_tokens 
-                True,  # thinking_mode
-                2000,  # thinking_budget_tokens
-                auto_execute_tools
-            )
-        
-        # Prepare the response
-        return UserResponse(
-            message=Message(role="assistant", content=content),
-            conversation_id=conversation_id,
-            tool_calls=tool_calls
-        )
-        
-    except Exception as e:
-        logger.error(f"Error processing tool results: {str(e)}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-@app.get("/api/tools")
-async def get_tools():
-    """
-    Get the list of available tools and their schemas.
-    """
-    return {"tools": TOOL_DEFINITIONS}
+# Include routers
+app.include_router(chat_router)
+app.include_router(conversation_router)
+app.include_router(files_router)
 
 # Add root route to redirect to frontend
 @app.get("/")
@@ -658,255 +69,6 @@ async def root():
 
 # Mount static files for frontend
 app.mount("/frontend", StaticFiles(directory=os.path.join(os.path.dirname(os.path.dirname(__file__)), "frontend")), name="frontend")
-
-# 添加一个新的路由来获取特定会话的消息
-@app.get("/api/conversation/{conversation_id}/messages")
-async def get_conversation_messages(conversation_id: str):
-    """
-    Get message history for a specific conversation.
-    Used for front-end polling to get results of automatic tool execution and new assistant responses.
-    """
-    logger.info(f"Getting messages for conversation {conversation_id}")
-    
-    if conversation_id not in conversations:
-        logger.warning(f"Conversation not found: {conversation_id}")
-        raise HTTPException(status_code=404, detail="Conversation not found")
-    
-    try:
-        # Get conversation history
-        history = conversations[conversation_id]
-        
-        # Check conversation status
-        # Find the last message, if it's assistant message and no tool calls, then consider conversation completed
-        status = "in_progress"
-        if history and len(history) > 0:
-            last_message = history[-1]
-            if last_message["role"] == "assistant":
-                # Check if there are tool calls
-                has_tool_call = False
-                for item in last_message["content"]:
-                    if isinstance(item, dict) and item.get("type") == "tool_use":
-                        has_tool_call = True
-                        break
-                
-                if not has_tool_call:
-                    status = "completed"
-        
-        # Get the conversation root directory if available
-        root_dir = conversation_root_dirs.get(conversation_id, None)
-        
-        return {
-            "conversation_id": conversation_id,
-            "messages": history,
-            "status": status,
-            "root_dir": root_dir
-        }
-        
-    except Exception as e:
-        logger.error(f"Error getting conversation messages: {str(e)}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-# 添加一个新的端点获取会话根目录
-@app.get("/api/conversation/{conversation_id}/root")
-async def get_conversation_root(conversation_id: str):
-    """
-    Get the root directory for a specific conversation.
-    """
-    try:
-        logger.info(f"Getting root directory for conversation {conversation_id}")
-        
-        if conversation_id not in conversation_root_dirs:
-            raise HTTPException(status_code=404, detail="Conversation root directory not found")
-        
-        root_dir = conversation_root_dirs[conversation_id]
-        
-        return {
-            "conversation_id": conversation_id,
-            "root_dir": root_dir
-        }
-        
-    except Exception as e:
-        logger.error(f"Error getting conversation root directory: {str(e)}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-# 添加一个新的端点用于取消自动工具执行
-@app.post("/api/conversation/{conversation_id}/cancel")
-async def cancel_auto_execution(conversation_id: str):
-    """
-    Cancel ongoing automatic tool execution.
-    Allows user to interrupt long-running automatic tool execution process.
-    """
-    try:
-        logger.info(f"Received cancel automatic execution request, conversation ID: {conversation_id}")
-        
-        if conversation_id not in conversations:
-            raise HTTPException(status_code=404, detail="Conversation not found")
-        
-        # Mark automatic execution task for this conversation as cancelled
-        auto_execute_tasks[conversation_id] = "cancelled"
-        
-        # Add system message to conversation history
-        if conversation_id in conversations:
-            conversations[conversation_id].append({
-                "role": "system", 
-                "content": [{"type": "text", "text": "Automatic tool execution cancelled by user"}]
-            })
-        
-        return {"status": "success", "message": "Automatic tool execution cancelled"}
-        
-    except Exception as e:
-        logger.error(f"Error cancelling automatic execution: {str(e)}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-# 添加一个新的端点用于提供会话目录中的文件
-@app.get("/api/conversation/{conversation_id}/files/{file_path:path}")
-async def serve_conversation_file(conversation_id: str, file_path: str):
-    """
-    Serve a file from a conversation directory
-    
-    Args:
-        conversation_id: Conversation ID
-        file_path: Path to the file within the conversation directory
-    """
-    try:
-        logger.info(f"[FILE SERVING] Request to serve file: {file_path} for conversation: {conversation_id}")
-        
-        # Check if we have a root directory for this conversation
-        if conversation_id not in conversation_root_dirs:
-            logger.error(f"[FILE SERVING] Root directory not found for conversation: {conversation_id}")
-            raise HTTPException(status_code=404, detail="Conversation not found")
-        
-        root_dir = conversation_root_dirs[conversation_id]
-        logger.info(f"[FILE SERVING] Found root directory: {root_dir}")
-        
-        # Resolve the full file path
-        full_path = os.path.join(root_dir, file_path)
-        logger.info(f"[FILE SERVING] Resolved full file path: {full_path}")
-        
-        # Check if the file exists
-        if not os.path.isfile(full_path):
-            logger.error(f"[FILE SERVING] File not found: {full_path}")
-            raise HTTPException(status_code=404, detail="File not found")
-        
-        # Get file size for logging
-        file_size = os.path.getsize(full_path)
-        logger.info(f"[FILE SERVING] File size: {file_size} bytes")
-        
-        # Determine content type based on file extension
-        file_extension = os.path.splitext(file_path)[1].lower()
-        content_type, _ = mimetypes.guess_type(file_path)
-        
-        # Use explicit content types for common file types to ensure proper handling
-        if file_extension == '.md':
-            content_type = 'text/markdown'
-        elif file_extension == '.png':
-            content_type = 'image/png'
-        elif file_extension in ['.jpg', '.jpeg']:
-            content_type = 'image/jpeg'
-        elif file_extension == '.html':
-            content_type = 'text/html'
-        
-        logger.info(f"[FILE SERVING] Determined content type: {content_type}")
-        
-        # Add additional headers for debugging
-        headers = {
-            "X-File-Path": file_path,
-            "X-File-Size": str(file_size),
-            "X-Content-Type": content_type or "unknown",
-            "X-Debug-Full-Path": full_path,
-        }
-        
-        # Ensure Cache-Control headers for proper browser caching
-        if content_type and content_type.startswith('image/'):
-            headers["Cache-Control"] = "public, max-age=3600"
-        
-        logger.info(f"[FILE SERVING] Serving file with headers: {headers}")
-        
-        # Return the file as a response
-        logger.info(f"[FILE SERVING] Successfully serving file: {file_path}")
-        return FileResponse(
-            path=full_path, 
-            filename=os.path.basename(file_path),
-            media_type=content_type,
-            headers=headers
-        )
-    except Exception as e:
-        logger.error(f"[FILE SERVING] Error serving file: {str(e)}")
-        import traceback
-        logger.error(f"[FILE SERVING] Traceback: {traceback.format_exc()}")
-        raise HTTPException(status_code=500, detail=f"Error serving file: {str(e)}")
-
-# 添加一个新的端点用于列出会话目录中的文件
-@app.get("/api/conversation/{conversation_id}/files")
-async def list_conversation_files(conversation_id: str):
-    """
-    List files in a conversation directory
-    
-    Args:
-        conversation_id: Conversation ID
-    """
-    try:
-        logger.info(f"[FILE LISTING] Request to list files for conversation: {conversation_id}")
-        
-        # Check if we have a root directory for this conversation
-        if conversation_id not in conversation_root_dirs:
-            logger.error(f"[FILE LISTING] Root directory not found for conversation: {conversation_id}")
-            raise HTTPException(status_code=404, detail="Conversation not found")
-        
-        root_dir = conversation_root_dirs[conversation_id]
-        logger.info(f"[FILE LISTING] Found root directory: {root_dir}")
-        
-        # Find all files in the directory and subdirectories
-        files = []
-        for dirpath, dirnames, filenames in os.walk(root_dir):
-            for filename in filenames:
-                # Skip temporary or hidden files
-                if filename.startswith('.') or filename.endswith('.tmp'):
-                    continue
-                
-                full_path = os.path.join(dirpath, filename)
-                rel_path = os.path.relpath(full_path, root_dir)
-                
-                # Determine the file type based on extension
-                file_extension = os.path.splitext(filename)[1].lower()
-                content_type, _ = mimetypes.guess_type(filename)
-                
-                # Get file size
-                file_size = os.path.getsize(full_path)
-                
-                # Determine file type category
-                file_type = "other"
-                if content_type:
-                    if content_type.startswith('image/'):
-                        file_type = "image"
-                    elif content_type.startswith('text/'):
-                        file_type = "text"
-                    elif content_type.startswith('application/'):
-                        file_type = "application"
-                
-                # Create file URL
-                file_url = f"/api/conversation/{conversation_id}/files/{rel_path}"
-                
-                logger.info(f"[FILE LISTING] Found file: {rel_path}, type: {file_type}, size: {file_size} bytes, content-type: {content_type}")
-                
-                files.append({
-                    "name": filename,
-                    "path": rel_path,
-                    "url": file_url,
-                    "size": file_size,
-                    "content_type": content_type,
-                    "type": file_type,
-                    "extension": file_extension,
-                })
-        
-        logger.info(f"[FILE LISTING] Found {len(files)} files in conversation {conversation_id}")
-        return {"files": files}
-        
-    except Exception as e:
-        logger.error(f"[FILE LISTING] Error listing files: {str(e)}")
-        import traceback
-        logger.error(f"[FILE LISTING] Traceback: {traceback.format_exc()}")
-        raise HTTPException(status_code=500, detail=f"Error listing files: {str(e)}")
 
 if __name__ == "__main__":
     import uvicorn
