+++ claude-tooling/app/api/models/__init__.py
@@ -1,0 +1,5 @@
+"""
+Models package for the API.
+"""
+
+from .schemas import Message, ToolUse, ToolOutput, UserRequest, UserResponse
