+++ claude-tooling/app/api/routes/__init__.py
@@ -1,0 +1,7 @@
+"""
+Routes package for the API.
+"""
+
+from .chat import router as chat_router
+from .conversation import router as conversation_router
+from .files import router as files_router
