+++ claude-tooling/app/tools/__init__.py
@@ -6,4 +6,8 @@ This package contains tools that can be called by Claude 3.7.
 from .tool_wrapper import TOOL_DEFINITIONS, process_tool_calls, format_tool_results_for_claude
 from .file_tools import save_file, read_file
 from .command_tools import run_command, install_python_package
-from .web_tools import search, extract_content 
+from .web_tools import search, extract_content
+
+"""
+Tools package for the API.
+"""
